Moved merge_sort back to Deck class. Simplicity. It passes the test.
