+++ lib/deck.rb
@@ -26,6 +26,43 @@ class Deck
     result
   end
 
-  include MergeSortable
+  def merge_sort(other = nil)
+    cards = self.cards
+    if other
+      result = []
+      (other.size + cards.size).times do
+        # require 'pry'; binding.pry
+        if cards.empty?
+          return result += other
+        elsif other.empty?
+          return result += cards
+        end
+
+        if cards[0].to_f < other[0].to_f
+          result << cards.shift
+        else
+          result << other.shift
+        end
+      end
+    else
+      result = cards.map{|el| [el]}
+
+
+      Math.log2(result.size).ceil.times do
+        new_result = []
+        result.each_slice(2).map do |a,b|
+          if b == nil
+            new_result << a
+          else
+            new_result << Deck.new(a).merge_sort(b)
+          end
+        end
+
+        result = new_result
+
+      end
+    end
+    result[0]
+  end
 
 end
